fix: make most style fixes to footer
