+++ components/header/BrownBanner.tsx
@@ -14,12 +14,18 @@ export const BrownBanner: React.FC = () => {
           </h1>
         </div>
       </div>
+
       <Button
         variant="outline_primary"
         className="absolute right-[49px] top-[50%] transform -translate-y-1/2 border-red-university text-red-university font-semibold rounded-none w-[159px] h-[50px] hover:bg-red-university hover:text-white hover:border-red-university"
       >
-        <MessagesSquareIcon className="mr-2 h-[26px] w-[26px]" />
-        Work with Us
+        <a 
+          href="mailto:ccv-support@brown.edu" 
+          className="flex items-center justify-center w-full h-full"
+        >
+          <MessagesSquareIcon className="mr-2 h-[26px] w-[26px]" />
+          Work with Us
+        </a>
       </Button>
     </header>
   );
